+++ src/types.ts
@@ -1,6 +1,6 @@
 /**
- * types.ts 10.1.2-dev
- * Copyright (c) 2021 Alain Dumesny - see GridStack root license
+ * types.ts 10.3.1-dev
+ * Copyright (c) 2021-2024 Alain Dumesny - see GridStack root license
  */
 
 import { GridStack } from './gridstack';
@@ -84,20 +84,23 @@ export type AddRemoveFcn = (parent: HTMLElement, w: GridStackWidget, add: boolea
 /** optional function called during save() to let the caller add additional custom data to the GridStackWidget structure that will get returned */
 export type SaveFcn = (node: GridStackNode, w: GridStackWidget) => void;
 
+/** optional function called during load()/addWidget() to let the caller create custom content other than plan text */
+export type RenderFcn = (el: HTMLElement, w: GridStackWidget) => void;
+
 export type ResizeToContentFcn = (el: GridItemHTMLElement) => void;
 
-/** describes the responsive nature of the grid */
+/** describes the responsive nature of the grid. NOTE: make sure to have correct extra CSS to support this. */
 export interface Responsive {
-  /** wanted width to maintain (+-50%) to dynamically pick a column count */
+  /** wanted width to maintain (+-50%) to dynamically pick a column count. NOTE: make sure to have correct extra CSS to support this. */
   columnWidth?: number;
-  /** maximum number of columns allowed (default: 12). Note: make sure to have correct extra CSS to support this.*/
+  /** maximum number of columns allowed (default: 12). NOTE: make sure to have correct extra CSS to support this. */
   columnMax?: number;
-  /** global re-layout mode when changing columns */
-  layout?: ColumnOptions;
+  /** explicit width:column breakpoints instead of automatic 'columnWidth'. NOTE: make sure to have correct extra CSS to support this. */
+  breakpoints?: Breakpoint[];
   /** specify if breakpoints are for window size or grid size (default:false = grid) */
   breakpointForWindow?: boolean;
-  /** explicit width:column breakpoints instead of automatic 'columnWidth'. Note: make sure to have correct extra CSS to support this.*/
-  breakpoints?: Breakpoint[];
+  /** global re-layout mode when changing columns */
+  layout?: ColumnOptions;
 }
 
 export interface Breakpoint {
@@ -254,7 +257,7 @@ export interface GridStackOptions {
 
   /**
    * if true turns grid to RTL. Possible values are true, false, 'auto' (default?: 'auto')
-   * See [example](http://gridstack.github.io/gridstack.js/demo/rtl.html)
+   * See [example](http://gridstack.github.io/gridstack.js/demo/right-to-left(rtl).html)
    */
   rtl?: boolean | 'auto';